@@ -1,8 +1,7 @@
-﻿namespace CalculatorLib
+﻿namespace CalculatorLib;
+
+public class Calculator
 {
-    public class Calculator
-    {
-        public double Add(double x, double y) => x + y;
-        public double Subtract(double x, double y) => x - y;
-    }
+    public double Add(double x, double y) => x + y;
+    public double Subtract(double x, double y) => x - y;
 }
