@@ -1,8 +1,7 @@
-﻿namespace DynamicSample
+﻿namespace DynamicSample;
+
+public class Person
 {
-    public class Person
-    {
-        public string? FirstName { get; set; }
-        public string? LastName { get; set; }
-    }
+    public string? FirstName { get; set; }
+    public string? LastName { get; set; }
 }
