@@ -1,7 +1,4 @@
-﻿using System;
-using DynamicFileReader;
-
-var employeeList = DynamicFileHelper.ParseFile("EmployeeList.txt");
+﻿var employeeList = DynamicFileHelper.ParseFile("EmployeeList.txt");
 foreach (var employee in employeeList)
 {
     Console.WriteLine($"{employee.FirstName} {employee.LastName} lives in {employee.City}, {employee.State}.");
