@@ -1,10 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Reflection;
-using System.Text;
-
-StringBuilder OutputText = new();
+﻿StringBuilder OutputText = new();
 
 // modify this line to retrieve details of any other data type
 Type t = typeof(double);
